systemd: refactor property parsers for int values in CurrentTasksCount, etc
As suggested by Maciej, Michael and Samuele in snapcore/snapd#10354.

+++ systemd/emulation.go
@@ -99,7 +99,7 @@ func (s *emulation) CurrentMemoryUsage(unit string) (quantity.Size, error) {
 	return 0, errNotImplemented
 }
 
-func (s *emulation) CurrentTasksCount(unit string) (int, error) {
+func (s *emulation) CurrentTasksCount(unit string) (uint64, error) {
 	return 0, errNotImplemented
 }
